+++ src/test/java/racing/RacingTest.java
@@ -1,15 +1,10 @@
 package racing;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -19,36 +14,25 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
  */
 class RacingTest {
 
-    class Racing {
-        private List<Car> cars;
-        private int attempts;
-
-        public Racing(Integer cars, Integer attempts) {
-            if (cars == null || cars.intValue() < 1) {
-                throw new IllegalArgumentException();
-            }
-
-            if (attempts == null || attempts.intValue() < 1) {
-                throw new IllegalArgumentException();
-            }
-            this.cars = createCars(cars);
-            this.attempts = attempts;
-        }
-
-        public void play() {
-
-        }
-
-        private List<Car> createCars(int count) {
-            return IntStream.range(0, count).mapToObj(i -> new Car()).collect(Collectors.toList());
-        }
-    }
-
     @Test
     @DisplayName(value = "생성")
     void create() {
         Racing racing = new Racing(1, 2);
-        assertThat(racing).isInstanceOf(Racing.class);
+        assertThat(racing.getCars().size()).isEqualTo(1);
+        assertThat(racing.getAttempts()).isEqualTo(2);
+    }
+
+    @RepeatedTest(name = "플레이", value = 10)
+    void play() {
+        Racing racing = new Racing(NumberHelper.getRandomValue(3) + 1, NumberHelper.getRandomValue(5) + 1);
+        List<Car> cars = racing.getCars();
+        int attempts = racing.getAttempts();
+
+        assertThat(cars.size()).isGreaterThanOrEqualTo(1);
+        assertThat(attempts).isGreaterThanOrEqualTo(1);
+
+        racing.play();
+        cars.forEach(car -> assertThat(car.getStep()).isGreaterThanOrEqualTo(0));
     }
 
     @Test
